check: Split the parallelized tree loader into a reusable component
The actual code change is minimal
